+++ odoo_score/odoo_score/scripts/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 from . import main
-from . import run_odoo_debug
+# from . import run_odoo_debug
 from . import rename_odoo_module
 from . import map_module_dependecies