+++ wok_code/wok_code/egg-info/history.rst
@@ -6,6 +6,10 @@
 * [IMP] deploy_odoo: new switches -l and -x
 * [NEW] do_git_checkout_new_branch.py
 * [IMP] do_migrate.py: new features
+* [IMP] run_odoo_debug.sh imported from odoo_score
+* [FIX] run_odoo_debug.sh: ODOO_COMMIT TEST not set when build template
+* [IMP] run_odoo_debug.sh: simulate server_wide_modules parameter for Odoo 7.0-
+
 
 2.0.5 (2023-01-13)
 ~~~~~~~~~~~~~~~~~~